Update routing configuration and enhance rate limiting functionality
- Changed the target for the root path from localhost:3000 to /events at localhost:4000 in config.yaml.
- Modified RateLimit struct in config.go to use int instead of int64 for rate and burst values.
- Updated getLimiter and RateLimitMiddleware functions to accept dynamic rate and burst parameters.
- Refactored HandleMuxRoutes to create reverse proxies based on the updated configuration and apply rate limiting accordingly.

+++ internal/config/config.go
@@ -19,8 +19,8 @@ type Route struct {
 }
 
 type RateLimit struct {
-	Rate  int64 `yaml:"rate"`
-	Burst int64 `yaml:"burst"`
+	Rate  int `yaml:"rate"`
+	Burst int `yaml:"burst"`
 }
 
 var ConfigFileData Config
